+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Unreleased
 
+[...]
+
+# v25.0.0 (02/04/2020)
+
+- **[BREAKING CHANGE]** Use `title` instead of `children` prop in `Stepper`
+- **[NEW]** Add `CalendarIcon` component
+- **[NEW]** Add `SearchForm` component
+
+# v24.1.1 (01/04/2020)
+
 - **[FIX]** Fix `Profile` medium size rendering
 - **[FIX]** Fix `TextDisplay1` scale
 - **[UPDATE]** Import `icons` as default instead of named export for a better code splitting
-- [...]
+- **[UPDATE]** Adjust `jsx-a11y` linter settings to recommended
 
 # v24.1.0 (25/03/2020)